Update rocky Dockerfiles
Add README describing how to run these Dockerfiles

* Dockerfile is used to verify upgrades, by default from 7.34.0 to 8.10.0.
* Dockerfile.versions is used to install multiple release packages and verify
  you can get a listing.

Update `Dockerfile` and `Dockerfile.versions` to allow credentials to be
specified, taking care not to save the credentials in the docker image.

Update `Dockerfile.versions` to download the puppet8 release package from
yum-support and after installation, configure the username and password in the
repo config.

Added `run-versions.shUpdate `Dockerfile` to copy `run-upgrade.sh` into the
docker image and run it using an externally provided password.

Update `run-upgrade.sh` to pass username and password as facts to `upgrade.pp`.
When the upgrade is performed, the module will create pc_repo with the provided
credentials and use that to upgrade to 8.10.0.

+++ docker/bin/upgrade.sh
@@ -8,23 +8,27 @@
 # - PLATFORM: The platform on which the upgrade should occur. This also
 #             supports comma-separated lists. Available:
 #             - `ubuntu`
-#             - `centos`
 #             - `rocky`
 #             Default: `ubuntu`
 # - BEFORE: The puppet-agent package version that is installed prior to upgrade.
-#           Default: 1.10.14
+#           Default: 7.34.0
 # - AFTER: The puppet-agent package version that should exist after upgrade.
-#          Default: 6.2.0
+#          Default: 8.1.0
 set -e
 
+if [ -z "${PUPPET_FORGE_TOKEN}" ]; then
+    echo "Environment variable PUPPET_FORGE_TOKEN must be set"
+    exit 1
+fi
+
 cd "$(dirname "$0")/../.."
-platforms=${1:-ubuntu}
+platforms=${1:-rocky}
 before=${2:-7.34.0}
 after=${3:-8.10.0}
 for platform in ${platforms//,/ }
 do
     docker build --rm -f docker/$platform/Dockerfile . -t pa-dev:$platform \
         --build-arg before=${before}
-    docker run --rm -ti pa-dev:$platform ${after}
+    docker run -e PUPPET_FORGE_TOKEN --rm -ti pa-dev:$platform ${after}
 done
 echo Complete
